v1.15
Further changes to make queue more memory-efficient. The older model is still available when template parameter Priority is set to plf::speed. The newer model, which is the default of Priority == plf::memory, reduces performance but decreases memory usage by up to 30%, depending on number of elements and size of the element type. Benchmarks are updated to show the results of both models.